backend/requirements.txt
litellm>=1.35.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
